refacto: rename base to base_image
