+++ cargo-green/src/base_image.rs
